remove config files, and fixes offline recon docs
