+++ docs/tutorial/reconstruction_system/make_fragments.rst
@@ -25,8 +25,8 @@ Register RGBD image pairs
 
 .. literalinclude:: ../../../examples/python/reconstruction_system/make_fragments.py
    :language: python
-   :lineno-start: 58
-   :lines: 27,59-88
+   :lineno-start: 46
+   :lines: 27,47-76
    :linenos:
 
 The function reads a pair of RGBD images and registers the ``source_rgbd_image``
@@ -46,8 +46,8 @@ Multiway registration
 
 .. literalinclude:: ../../../examples/python/reconstruction_system/make_fragments.py
    :language: python
-   :lineno-start: 88
-   :lines: 27,89-136
+   :lineno-start: 76
+   :lines: 27,77-123
    :linenos:
 
 This script uses the technique demonstrated in
@@ -63,7 +63,7 @@ function ``optimize_posegraph_for_fragment``.
 .. literalinclude:: ../../../examples/python/reconstruction_system/optimize_posegraph.py
    :language: python
    :lineno-start: 51
-   :lines: 27,52-64
+   :lines: 27,52-63
    :linenos:
 
 This function calls ``global_optimization`` to estimate poses of the RGBD images.
@@ -75,8 +75,8 @@ Make a fragment
 
 .. literalinclude:: ../../../examples/python/reconstruction_system/make_fragments.py
    :language: python
-   :lineno-start: 136
-   :lines: 27,137-158
+   :lineno-start: 124
+   :lines: 27,125-146
    :linenos:
 
 Once the poses are estimates, :ref:`/tutorial/pipelines/rgbd_integration.ipynb`
@@ -87,8 +87,8 @@ Batch processing
 
 .. literalinclude:: ../../../examples/python/reconstruction_system/make_fragments.py
    :language: python
-   :lineno-start: 193
-   :lines: 27,194-217
+   :lineno-start: 181
+   :lines: 27,182-205
    :linenos:
 
 The main function calls each individual function explained above.
