--- examples/python/reconstruction_system/config/indoor_lidar_rgbd/lobby.json
@@ -1,14 +1,0 @@
-{
-    "name": "Indoor Lidar-RGBD Dataset: http://redwood-data.org/indoor_lidar_rgbd/download.html",
-    "path_dataset": "dataset/indoor_lidar_rgbd/lobby/",
-    "path_intrinsic": "",
-    "max_depth": 3.0,
-    "voxel_size": 0.05,
-    "max_depth_diff": 0.07,
-    "preference_loop_closure_odometry": 0.1,
-    "preference_loop_closure_registration": 5.0,
-    "tsdf_cubic_size": 3.0,
-    "icp_method": "color",
-    "global_registration": "ransac",
-    "python_multi_threading": true
-}
