--- examples/python/reconstruction_system/config/redwood_objects/plant.json
@@ -1,14 +1,0 @@
-{
-    "name": "Single object scans http://redwood-data.org/3dscan/index.html",
-    "path_dataset": "dataset/redwood_objects/plant/",
-    "path_intrinsic": "",
-    "max_depth": 2.0,
-    "voxel_size": 0.02,
-    "max_depth_diff": 0.03,
-    "preference_loop_closure_odometry": 0.1,
-    "preference_loop_closure_registration": 5.0,
-    "tsdf_cubic_size": 2.0,
-    "icp_method": "color",
-    "global_registration": "ransac",
-    "python_multi_threading": true
-}
