--- examples/python/reconstruction_system/config/redwood_simulated/livingroom1-clean.json
@@ -1,23 +1,0 @@
-{
-    "name": "Redwood Dataset: http://redwood-data.org/indoor/dataset.html",
-    "path_dataset": "dataset/redwood_simulated/livingroom1-clean/",
-    "path_intrinsic": "",
-    "max_depth": 5.0,
-    "voxel_size": 0.05,
-    "max_depth_diff": 0.07,
-    "preference_loop_closure_odometry": 0.1,
-    "preference_loop_closure_registration": 5.0,
-    "tsdf_cubic_size": 4.0,
-    "icp_method": "point_to_point",
-    "global_registration": "ransac",
-    "python_multi_threading": true,
-    "max_iterations": 5,
-    "sdf_trunc": 0.04,
-    "block_count": 40000,
-    "distance_threshold": 0.07,
-    "fitness_threshold": 0.3,
-    "regularizer_weight": 1,
-    "method": "slac",
-    "device": "CPU:0",
-    "save_output_as": "pointcloud" 
-}
