--- examples/python/reconstruction_system/config/redwood_simulated/office1-simulated.json
@@ -1,15 +1,0 @@
-{
-    "name": "Redwood Dataset: http://redwood-data.org/indoor/dataset.html",
-    "path_dataset": "dataset/redwood_simulated/office1-simulated/",
-    "path_intrinsic": "",
-    "n_frames_per_fragment": 50,
-    "max_depth": 5.0,
-    "voxel_size": 0.05,
-    "max_depth_diff": 0.07,
-    "preference_loop_closure_odometry": 0.1,
-    "preference_loop_closure_registration": 20.0,
-    "tsdf_cubic_size": 4.0,
-    "icp_method": "point_to_point",
-    "global_registration": "ransac",
-    "python_multi_threading": true
-}
