--- examples/python/reconstruction_system/config/stanford/cactusgarden.json
@@ -1,14 +1,0 @@
-{
-    "name": "Open3D reconstruction tutorial http://open3d.org/docs/release/tutorial/reconstruction_system/system_overview.html",
-    "path_dataset": "dataset/stanford/cactusgarden/",
-    "path_intrinsic": "",
-    "max_depth": 3.0,
-    "voxel_size": 0.05,
-    "max_depth_diff": 0.07,
-    "preference_loop_closure_odometry": 0.1,
-    "preference_loop_closure_registration": 5.0,
-    "tsdf_cubic_size": 3.0,
-    "icp_method": "color",
-    "global_registration": "ransac",
-    "python_multi_threading": true
-}
