+++ examples/python/reconstruction_system/make_fragments.py
@@ -34,7 +34,7 @@ import open3d as o3d
 pyexample_path = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 sys.path.append(pyexample_path)
 
-from open3d_example import join, make_clean_folder, get_rgbd_file_lists, initialize_opencv
+from open3d_example import *
 
 sys.path.append(os.path.dirname(os.path.abspath(__file__)))
 from optimize_posegraph import optimize_posegraph_for_fragment
@@ -43,18 +43,6 @@ from optimize_posegraph import optimize_posegraph_for_fragment
 with_opencv = initialize_opencv()
 if with_opencv:
     from opencv_pose_estimation import pose_estimation
-
-
-def read_rgbd_image(color_file, depth_file, convert_rgb_to_intensity, config):
-    color = o3d.io.read_image(color_file)
-    depth = o3d.io.read_image(depth_file)
-    rgbd_image = o3d.geometry.RGBDImage.create_from_color_and_depth(
-        color,
-        depth,
-        depth_scale=config["depth_scale"],
-        depth_trunc=config["max_depth"],
-        convert_rgb_to_intensity=convert_rgb_to_intensity)
-    return rgbd_image
 
 
 def register_one_rgbd_pair(s, t, color_files, depth_files, intrinsic,
